fix: rename `toggle` folders to `switch`
