+++ src/lib/components/switch/QSwitch.scss
