+++ src/lib/components/switch/docs.ts
