+++ src/lib/components/switch/props.ts
